플레이어 싱글톤 컴포넌트 추가/ 메뉴 Scene 제작중

+++ Assets/Scripts/Creature/Player/PlayerController.cs
@@ -195,7 +195,7 @@ public class PlayerController : CreatureController
         }
     }
 
-    public void OnHit(int damage, Vector2 knockback)
+    public void OnHit(float damage, Vector2 knockback)
     {
         rb.linearVelocity = new Vector2(knockback.x, rb.linearVelocityY + knockback.y);
     }
